+++ Common/src/main/java/corgitaco/enhancedcelestials/core/ECBlocks.java
@@ -1,25 +1,28 @@
 package corgitaco.enhancedcelestials.core;
 
-import corgitaco.enhancedcelestials.EnhancedCelestials;
 import corgitaco.enhancedcelestials.block.SpaceMossBlock;
 import corgitaco.enhancedcelestials.block.SpaceMossCarpetBlock;
 import corgitaco.enhancedcelestials.block.SpaceMossGrassBlock;
-import corgitaco.enhancedcelestials.reg.RegistrationProvider;
-import corgitaco.enhancedcelestials.reg.RegistryObject;
-import net.minecraft.core.registries.Registries;
+import corgitaco.enhancedcelestials.platform.services.RegistrationService;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 
-public final class ECBlocks {
-    private static final RegistrationProvider<Block> BLOCKS = RegistrationProvider.get(Registries.BLOCK, EnhancedCelestials.MOD_ID);
+import java.util.function.Supplier;
 
-    public static final RegistryObject<Block> METEOR = BLOCKS.register("meteor", () -> new Block(BlockBehaviour.Properties.copy(Blocks.STONE)));
-    public static final RegistryObject<Block> SPACE_MOSS_CARPET = BLOCKS.register("space_moss_carpet", () -> new SpaceMossCarpetBlock(BlockBehaviour.Properties.copy(Blocks.MOSS_CARPET)));
-    public static final RegistryObject<Block> SPACE_MOSS_BLOCK = BLOCKS.register("space_moss_block", () -> new SpaceMossBlock(BlockBehaviour.Properties.copy(Blocks.MOSS_BLOCK).randomTicks()));
-    public static final RegistryObject<Block> SPACE_MOSS_GRASS = BLOCKS.register("space_moss_grass", () -> new SpaceMossGrassBlock(BlockBehaviour.Properties.copy(Blocks.GRASS)));
+public final class ECBlocks {
+
+    public static final Supplier<Block> METEOR = register("meteor", () -> new Block(BlockBehaviour.Properties.copy(Blocks.STONE)));
+    public static final Supplier<Block> SPACE_MOSS_CARPET = register("space_moss_carpet", () -> new SpaceMossCarpetBlock(BlockBehaviour.Properties.copy(Blocks.MOSS_CARPET)));
+    public static final Supplier<Block> SPACE_MOSS_BLOCK = register("space_moss_block", () -> new SpaceMossBlock(BlockBehaviour.Properties.copy(Blocks.MOSS_BLOCK).randomTicks()));
+    public static final Supplier<Block> SPACE_MOSS_GRASS = register("space_moss_grass", () -> new SpaceMossGrassBlock(BlockBehaviour.Properties.copy(Blocks.GRASS)));
 
     private ECBlocks() {
+    }
+
+    public static <B extends Block> Supplier<B> register(String id, Supplier<B> block) {
+        return RegistrationService.INSTANCE.register((net.minecraft.core.Registry<B>) BuiltInRegistries.BLOCK, id, block);
     }
 
     public static void classLoad() {