+++ Common/src/main/java/corgitaco/enhancedcelestials/core/ECSounds.java
@@ -1,23 +1,24 @@
 package corgitaco.enhancedcelestials.core;
 
 import corgitaco.enhancedcelestials.EnhancedCelestials;
-import corgitaco.enhancedcelestials.reg.RegistrationProvider;
-import corgitaco.enhancedcelestials.reg.RegistryObject;
+import corgitaco.enhancedcelestials.platform.services.RegistrationService;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
 
+import java.util.function.Supplier;
+
 public class ECSounds {
 
-    public static final RegistrationProvider<SoundEvent> SOUNDS = RegistrationProvider.get(Registries.SOUND_EVENT, EnhancedCelestials.MOD_ID);
+    public static final Supplier<SoundEvent> BLOOD_MOON = createSound("blood_moon");
+    public static final Supplier<SoundEvent> BLUE_MOON = createSound("blue_moon");
+    public static final Supplier<SoundEvent> HARVEST_MOON = createSound("harvest_moon");
 
-    public static final RegistryObject<SoundEvent> BLOOD_MOON = createSound("blood_moon");
-    public static final RegistryObject<SoundEvent> BLUE_MOON = createSound("blue_moon");
-    public static final RegistryObject<SoundEvent> HARVEST_MOON = createSound("harvest_moon");
-
-    public static RegistryObject<SoundEvent> createSound(String location) {
+    public static Supplier<SoundEvent> createSound(String location) {
         final ResourceLocation soundLocation = new ResourceLocation(EnhancedCelestials.MOD_ID, location);
-        return SOUNDS.register(location, () -> SoundEvent.createVariableRangeEvent(soundLocation));
+        return RegistrationService.INSTANCE.register(BuiltInRegistries.SOUND_EVENT, location, () -> SoundEvent.createVariableRangeEvent(soundLocation));
     }
 
+    public static void loadClass() {}
 }