+++ Common/src/main/java/corgitaco/enhancedcelestials/platform/Services.java
@@ -1,13 +1,11 @@
 package corgitaco.enhancedcelestials.platform;
 
 import corgitaco.enhancedcelestials.EnhancedCelestials;
-import corgitaco.enhancedcelestials.platform.services.IPlatformHelper;
 
 import java.util.ServiceLoader;
 
 public class Services {
 
-    public static final IPlatformHelper PLATFORM = load(IPlatformHelper.class);
 
     public static <T> T load(Class<T> clazz) {