+++ Common/src/main/java/corgitaco/enhancedcelestials/world/level/levelgen/structure/ECStructurePieceTypes.java
@@ -1,20 +1,18 @@
 package corgitaco.enhancedcelestials.world.level.levelgen.structure;
 
-import corgitaco.enhancedcelestials.EnhancedCelestials;
-import corgitaco.enhancedcelestials.reg.RegistrationProvider;
-import corgitaco.enhancedcelestials.reg.RegistryObject;
+import corgitaco.enhancedcelestials.platform.services.RegistrationService;
 import corgitaco.enhancedcelestials.world.level.levelgen.structure.crater.CraterPiece;
-import net.minecraft.core.registries.Registries;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.level.levelgen.structure.pieces.StructurePieceType;
 
 import java.util.Locale;
+import java.util.function.Supplier;
 
 public class ECStructurePieceTypes {
-    public static final RegistrationProvider<StructurePieceType> PROVIDER = RegistrationProvider.get(Registries.STRUCTURE_PIECE, EnhancedCelestials.MOD_ID);
-    public static final RegistryObject<StructurePieceType> CRATER_PIECE = setFullContextPieceId(CraterPiece::new, "CraterPiece");
+    public static final Supplier<StructurePieceType> CRATER_PIECE = setFullContextPieceId(() -> CraterPiece::new, "CraterPiece");
 
-    private static RegistryObject<StructurePieceType> setFullContextPieceId(StructurePieceType type, String id) {
-        return PROVIDER.register(id.toLowerCase(Locale.ROOT), () -> type);
+    private static Supplier<StructurePieceType> setFullContextPieceId(Supplier<StructurePieceType> type, String id) {
+        return RegistrationService.INSTANCE.register(BuiltInRegistries.STRUCTURE_PIECE, id.toLowerCase(Locale.ROOT), type);
     }
 
     public static void bootStrap() {