+++ Common/src/main/java/corgitaco/enhancedcelestials/world/level/levelgen/structure/ECStructureTypes.java
@@ -1,24 +1,22 @@
 package corgitaco.enhancedcelestials.world.level.levelgen.structure;
 
-import com.mojang.serialization.Codec;
-import corgitaco.enhancedcelestials.EnhancedCelestials;
-import corgitaco.enhancedcelestials.reg.RegistrationProvider;
-import corgitaco.enhancedcelestials.reg.RegistryObject;
+import corgitaco.enhancedcelestials.platform.services.RegistrationService;
 import corgitaco.enhancedcelestials.world.level.levelgen.structure.crater.CraterStructure;
-import net.minecraft.core.registries.Registries;
+import net.minecraft.core.Registry;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.level.levelgen.structure.Structure;
 import net.minecraft.world.level.levelgen.structure.StructureType;
 
+import java.util.Locale;
 import java.util.function.Supplier;
 
 public class ECStructureTypes {
 
-    public static final RegistrationProvider<StructureType<?>> PROVIDER = RegistrationProvider.get(Registries.STRUCTURE_TYPE, EnhancedCelestials.MOD_ID);
 
-    public static final RegistryObject<StructureType<CraterStructure>> CRATER = register("crater", () -> CraterStructure.CODEC);
+    public static final Supplier<StructureType<CraterStructure>> CRATER = register("crater", () -> () -> CraterStructure.CODEC);
 
-    private static <S extends Structure> RegistryObject<StructureType<S>> register(String id, Supplier<? extends Codec<S>> codec) {
-        return PROVIDER.register(id, () -> codec::get);
+    private static <S extends Structure, ST extends StructureType<S>> Supplier<ST> register(String id, Supplier<ST> type) {
+        return RegistrationService.INSTANCE.register((Registry<ST>) BuiltInRegistries.STRUCTURE_TYPE, id.toLowerCase(Locale.ROOT), type);
     }
 
     public static void loadClass() {