+++ Fabric/src/main/java/corgitaco/enhancedcelestials/fabric/platform/FabricPlatformHelper.java
@@ -1,0 +1,39 @@
+package corgitaco.enhancedcelestials.fabric.platform;
+
+import com.google.auto.service.AutoService;
+import corgitaco.enhancedcelestials.fabric.network.FabricNetworkHandler;
+import corgitaco.enhancedcelestials.network.S2CPacket;
+import corgitaco.enhancedcelestials.platform.services.IPlatformHelper;
+import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.server.level.ServerPlayer;
+
+import java.nio.file.Path;
+
+@AutoService(IPlatformHelper.class)
+public class FabricPlatformHelper implements IPlatformHelper {
+
+    @Override
+    public String getPlatformName() {
+        return "Fabric";
+    }
+
+    @Override
+    public boolean isModLoaded(String modId) {
+        return FabricLoader.getInstance().isModLoaded(modId);
+    }
+
+    @Override
+    public boolean isDevelopmentEnvironment() {
+        return FabricLoader.getInstance().isDevelopmentEnvironment();
+    }
+
+    @Override
+    public <P extends S2CPacket> void sendToClient(ServerPlayer player, P packet) {
+        FabricNetworkHandler.sendToPlayer(player, packet);
+    }
+
+    @Override
+    public Path configDir() {
+        return FabricLoader.getInstance().getConfigDir();
+    }
+}